+++ lib/enums/UrlEnum.ts
@@ -18,4 +18,6 @@ export enum UrlEnum {
   ATTRIBUTE_CHECK = '/check-attribute',
   GET_USER_DATA = '/get-user-details',
   BATCH_EVENTS = '/server-side/batch-events-v2',
+  SET_ALIAS = '/user-alias/setUserAlias',
+  GET_ALIAS = '/user-alias/getAliasUserId',
 }